Refactor JWT claim extraction
